+++ app-bazlang/src/main/java/com/davidconneely/bazlang/InterpreterReplHandler.java
@@ -108,7 +108,8 @@ public final class InterpreterReplHandler implements ReplHandler {
     } else if (ctx instanceof BazLangParser.EditCmdContext edit) {
       final int lineNum = (int) executor.evalNum(edit.numExpr());
       if (lineNum < Limits.MIN_LINE_LABEL || lineNum > Limits.MAX_LINE_LABEL) {
-        throw new ReportException(ReportCode.INTEGER_OUT_OF_RANGE, 0, "Line number out of range");
+        throw new ReportException(
+            ReportCode.INTEGER_OUT_OF_RANGE, 0, 1, "Line number out of range");
       }
       final var programLine = state.program().get(lineNum);
       if (input != null) {